Software Engineering/Cloud & Fog Computing/labs/Lab3/cloud_dashboard/requirements.txt
paho-mqtt==2.1.0
Flask==3.1.2
